requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
wandb>=0.16.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pyyaml>=6.0
tqdm>=4.66.0
huggingface-hub>=0.19.0
jupyter>=1.0.0